refactor: migrate 4 more production files from complexity_classifier to canonical imports
Migrated remaining production files to use canonical imports:
- orchestrator.py: ComplexityClassifier → victor.framework.task.TaskComplexityService
- orchestrator_factory.py: ComplexityClassifier → victor.framework.task.TaskComplexityService
- service_provider.py: ComplexityClassifier → victor.framework.task.TaskComplexityService
- chat.py: ComplexityClassifier → victor.framework.task.TaskComplexityService

Total: 6/6 production files migrated from deprecated complexity_classifier shim

Part of Phase 2: complexity_classifier shim removal
